requirements.txt
streamlit==1.46.1
langchain==0.1.0
langchain-community==0.0.16
langchain-openai==0.0.5
faiss-cpu==1.7.4
openai==1.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.26.2
scipy==1.11.4
sounddevice==0.4.6
requests==2.31.0
uuid==1.30
